+++ simpleble/include/simpleble/Types.h
@@ -15,11 +15,6 @@ using BluetoothUUID = std::string;
 // and pointers to uint8_t.
 using ByteArray = std::string;
 
-struct BluetoothService {
-    BluetoothUUID uuid;
-    std::vector<BluetoothUUID> characteristics;
-};
-
 enum class OperatingSystem {
     WINDOWS,
     MACOS,